+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 - Paraguay service stations are added to the map.
 - Add clustering in the service stations.
 - The sidebar is added to show information about the service stations.
+- Add group layers.